refactor(edgeless): split components.ts to multiple files
